// src/lib/api/generated/models/eventDelivery.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.1
 */

export interface EventDelivery {
  attempts: number;
  /** @nullable */
  claim_token?: string | null;
  created_at: string;
  event_id: number;
  id: number;
  /** @nullable */
  last_error?: string | null;
  /** @nullable */
  locked_until?: string | null;
  next_attempt_at: string;
  status: string;
  subscription_id: number;
  updated_at: string;
}